[REFACTOR] moved modified peptide generator to chemistry folder
