+++ src/openms/include/OpenMS/ANALYSIS/RNPXL/sources.cmake
@@ -4,7 +4,6 @@ set(directory include/OpenMS/ANALYSIS/RNPXL)
 ### list all header files of the directory here
 set(sources_list_h
 HyperScore.h
-ModifiedPeptideGenerator.h
 MorpheusScore.h
 PScore.h
 RNPxlFragmentAnnotationHelper.h
