+++ src/openms/include/OpenMS/CHEMISTRY/sources.cmake
@@ -19,6 +19,7 @@ DigestionEnzymeDB.h
 ModificationDefinition.h
 ModificationDefinitionsSet.h
 ModifiedNASequenceGenerator.h
+ModifiedPeptideGenerator.h
 ModificationsDB.h
 NASequence.h
 NucleicAcidSpectrumGenerator.h
