+++ src/openms/source/ANALYSIS/RNPXL/sources.cmake
@@ -4,7 +4,6 @@ set(directory source/ANALYSIS/RNPXL)
 ### list all filenames of the directory here
 set(sources_list
 HyperScore.cpp
-ModifiedPeptideGenerator.cpp
 MorpheusScore.cpp
 PScore.cpp
 RNPxlFragmentAnnotationHelper.cpp
